layout: Remove dot from enumerate
This has never worke in the past for me, but I finally got it working!
It took me forever to realise that this wasn't enumitem/latex's fault,
but was instead pandoc meddling with list lables/spacing. I have added
a regex to remove this code in my conversion/compilation script in my
dotfiles, which you can find here:
https://github.com/aramisgithub/dotfiles/blob/d7daaaf3f3124b00dcddf936f11069c73d2de090/home/bin/bin/convertit#L25